Task1/requirements-training.txt
# Training Requirements for NorgesGruppen Object Detection
# Install with: pip install -r requirements-training.txt

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Object Detection
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# COCO Tools
pycocotools>=2.0.6